Realign tests by runtime profile and semantics
