+++ Tests/CMakeLists.txt
@@ -41,6 +41,7 @@ endfunction()
 # Unit tests
 # -----------------------------------------------------------------------------
 set(GLAB_UNIT_TEST_SOURCES
+    Unit/TestAssetPath.cpp
     Unit/TestBuiltinTraits.cpp
     Unit/TestCamera.cpp
     Unit/TestCookedCatalog.cpp
@@ -77,9 +78,12 @@ glab_add_test_executable(rtrlab_unit_tests "Unit." "unit;common"
 # Integration tests
 # -----------------------------------------------------------------------------
 set(GLAB_INTEGRATION_TEST_SOURCES
+    Integration/TestCookPipeline.cpp
     Integration/TestLoggerConsoleSink.cpp
     Integration/TestMountBackend.cpp
     Integration/TestInputRecordingIO.cpp
+    Integration/TestPackagePipeline.cpp
+    Integration/TestSourceIndexPipeline.cpp
 )
 
 glab_add_test_executable(rtrlab_integration_tests "Integration." "integration;common"
@@ -87,8 +91,8 @@ glab_add_test_executable(rtrlab_integration_tests "Integration." "integration;co
 )
 
 set(GLAB_DEV_PROFILE_TEST_SOURCES
-    Integration/TestMountBackendDev.cpp
-    Integration/TestSourceCatalogDev.cpp
+    Integration/TestCatalogRegistryDev.cpp
+    Integration/TestMountDiscoveryDev.cpp
     Integration/TestRootDiscoveryDev.cpp
 )
 
@@ -97,6 +101,7 @@ glab_add_test_executable(rtrlab_dev_profile_tests "DevProfile." "integration;dev
 )
 
 set(GLAB_SHIPPING_PROFILE_TEST_SOURCES
+    Integration/TestMountDiscoveryShipping.cpp
     Integration/TestRootDiscoveryShipping.cpp
 )
