+++ Tests/Unit/TestBuiltinTraits.cpp
@@ -142,48 +142,6 @@ TEST(BuiltinTraitsTests, AssetPathRoundTrip)
     EXPECT_EQ(output.String(), "/Project/Textures/Grassy_Square");
 }
 
-TEST(BuiltinTraitsTests, AssetPathTryCreateAcceptsContentPathsWithoutExtensions)
-{
-    EXPECT_TRUE(Resource::AssetPath::TryCreate("/Project/Textures/Grassy_Square").has_value());
-    EXPECT_TRUE(Resource::AssetPath::TryCreate("/Engine/Defaults/Materials/ErrorMaterial").has_value());
-}
-
-TEST(BuiltinTraitsTests, AssetPathTryCreateRejectsConfigDocumentPaths)
-{
-    EXPECT_FALSE(Resource::AssetPath::TryCreate("/Project/Config/Graphics.json").has_value());
-    EXPECT_FALSE(Resource::AssetPath::TryCreate("/Engine/Config/Graphics.json").has_value());
-}
-
-TEST(BuiltinTraitsTests, AssetPathDeserializeRejectsAbsoluteFilesystemPathAndLeavesOutputUnchanged)
-{
-    const auto original = Resource::AssetPath::TryCreate("/Project/Textures/Original");
-    ASSERT_TRUE(original.has_value());
-    Resource::AssetPath output = *original;
-
-    EXPECT_FALSE(Deserialize(PropertyTree("C:/Users/name/dev/RTRLab/Project/textures/Grassy_Square.jpg"), output));
-    EXPECT_EQ(output, *original);
-}
-
-TEST(BuiltinTraitsTests, AssetPathDeserializeRejectsSavedDomainAndLeavesOutputUnchanged)
-{
-    const auto original = Resource::AssetPath::TryCreate("/Project/Textures/Original");
-    ASSERT_TRUE(original.has_value());
-    Resource::AssetPath output = *original;
-
-    EXPECT_FALSE(Deserialize(PropertyTree("/Saved/Config/Input/DebugCameraControl.json"), output));
-    EXPECT_EQ(output, *original);
-}
-
-TEST(BuiltinTraitsTests, AssetPathDeserializeRejectsDocumentPathAndLeavesOutputUnchanged)
-{
-    const auto original = Resource::AssetPath::TryCreate("/Project/Textures/Original");
-    ASSERT_TRUE(original.has_value());
-    Resource::AssetPath output = *original;
-
-    EXPECT_FALSE(Deserialize(PropertyTree("/Project/Config/input/DebugCameraControl.json"), output));
-    EXPECT_EQ(output, *original);
-}
-
 TEST(BuiltinTraitsTests, Uint8RoundTrip)
 {
     PropertyTree tree;
